separated functionality for socket server and emitter
